moved utility method for Transform into Utils
