+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Transform/DCL.Components.Transform.asmdef
@@ -8,7 +8,8 @@
         "GUID:fbcc413e192ef9048811d47ab0aca0c0",
         "GUID:a881b57670b1d2747a6d7f9e32b63230",
         "GUID:5289a39f46259954ca74010d7c9ded16",
-        "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
+        "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
+        "GUID:b1087c5731ff68448a0a9c625bb7e52d"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
